BE/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "leeSwallow", email = "lmh2157@outlook.kr" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pgvector>=0.4.2",
    "protobuf>=6.33.0",
    "pydantic-settings>=2.13.1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[project.scripts]
be = "be:main"
be-grpc = "be.grpc_app:main"

[build-system]
requires = ["uv_build>=0.8.16,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
